Change the way cluster are referenced.
Instead of a 8 pointer, we use a 3B size + 5B pointer.
The size being the size of the cluster header.

On the cluster side, we don't need to store the cluster data.
But we need to store the size of the raw data.

To simplify the writing process, the "header" is not put at the end of
the data, thus one can write the data on the file and put the offset after.