Change the definition of ContentAddress key.
Instead of having the keydata being a boolean flag telling if we must
patch the following Content, we now store a uint telling how many
following content we should patch (chaining available).

We must not store a extra ContentAddress following the "patch"
contentaddress. As we store how many contentAddress we must patch, we know
the size of the key.