Change repartition of SizedOffset to 2/6 bytes.
The number of blob in a cluster is 4096.
Even with the maximum size to store the data in the cluster, the
cluster header size would be at maximum 4+8*(4096+2) = 32788.
This size can be stored in a U16, so we can use a SizedOffset 2/6 for the
cluster localization. The size of a contentPack is now limited to 256TB.

On the IndexedKeyStore, with a size of 65535 bytes, we can store from
8190 keys (if keysize is 8) to around 21800 usable entries.
This is far enough. Huge content (keysize==8) should go in real content
and if there is a lot of real different key, we can store the keys in
a plain key store.