Move the comparaison of property outside of `Finder`.
The finder criteria my be more complex than simply compare the value
in a entry with a reference entry.
This is especially the case with arx which now store the basename and the
extension in two different property.

It is better to let `Finder::find` get a `Comparator` object.
And it allow us to remove the resolver from `Finder`.