Avoid cloning `Arc` all the time.
Having static lifetime Reader and Stream is nice but it means that we
need to clone the `Rc<dyn Source>` all the time.
This is a pitty as most of the time, we create stream just as a
intermediate structure to help parsing.

This commit introduce `SubReader` and `Flux` which does the same thing
than `Reader` and `Stream` but borrow the `Arc` instead of cloning it.

The `Producable::produce` method now take a `Flux` instead of a `Stream`.
Methods `create_stream_foo` have also been rename to `create_flux_foo`.