Make the Source responsible to read and convert value.
Before that, Source was responsible to get a array and the
reader/sub_reader/stream/flux to convert to a value.

But getting a array is inefficient as we need to copy the array when we
could simply use a slice.
But File is a Source and we cannot return a slice pointing to a file.
By moving the convertion into the Source itself, Source can do the best
for the situation.

We also introduce a MemorySource and MemoryReader to wrap Source for
which we can get a real slice (so not File)