Keep a reference to value_store in the extend and remove Resolver.
With the change using a builder to build the raw_value, we can now
have advanced builder (in contrast to create the raw_value from the
layout::property).

In such, we can resolve the value_store_idx to value_store at creation of
the builder and make the builder directly create extend containing a
Rc<ValueStore>. This simplify a lot as we don't have to resolve
value_store_idx all the time. And so we can pass the ValueStorage only
at builder creation.

We don't need now a `Resolver` as it  was use to make the link between
a ValueStorage and the value_store_id in the extend.