Fix comparison of `creator::Value::Array`.
This is valid for different values to point to the same id in the value
store. This can happen when one is the prefix of the other.
This is especially true when all the value can be stored in the
fixed_array, in this case the id is 0 and so it points to "another" value.
We must compare the size in this case.