Do not keep the value_store and entry_store sorted at creation.
Keeping them sorted take too long. It is better to keep them unsorted
and sort them at the end.

This is possible as we use Vow everywhere and we can keep track of the
value and idx even after sorting it.

The sorting was use to identify duplicate values. We now do this two ways:
- For indexed value store, do a linear search, assuming the number of
  values will be limited
- For plain value store, store all values and do a filtering at the end,
  once the store is sorted.