Remove BufferedFile.
BufferedFile was buggy if we were trying to get a slice crossing the
boundary of the buffer (ie, offset < 512 and end > 512 for example).

This was used to avoid a copy in the buffer to read from the inner
buffer instead of copying to a `buf[size]`. But it is better to have
worst performance than buggy feature.