Use stdlib optimization about copying file on linux when possible.
The standard library has a nice optimisation on linux where it use
`copy_file_range` internally when we copy from one file to another one.
This avoid reading the data in user space to write it just after.

However, to do this, the specialization is based on the exact type passed
to `std::io::copy`. So we cannot rely on `dyn Read`/`dyn Write` to trigger
it.

If workaround this by adding a `get_file_source` in `InputReader` which
may return a `Take<File>`.

Using it, and so using the optimization show us x2 performance gain when
copying (part of) file in a uncompressed cluster.