Switch to `xz2` crate instead of `rust-lzma` for lzma encoding.
`rust-lzma` works well but it is based on a "old" version of liblzma
which in turn force a "old" version of "xz2", and so "lzma-sys" which
doesn't compile on Windows.

`tar2arx` wants to use `niffler` crate which is based on `xz2` and so,
`tar2arx` doesn't compile on Windows.
By moving to `xz2` (which seems better maintained/used) we fix `tar2arx`
issue.