Separate parsing from reading.
The main change is now that Source/Reader/Flux do not parse, they only read data.

It is the parser which get a slice from Source and then parse it.
This potentially create a performance dropdown as `File` has to allocate a
buffer to copy the bytes of our data and then allow Parser to parse.
Before that, `File` was directly read data in the data to parse (and then
parse (byteswap) inplace).

This will be mitigate in next commit by forcing read in memory the whole structure
first before parsing.