Do not mmap with `MADV_POPULATE_READ`.
`MADV_POPULATE_READ` is a bit too strong for what we need.
It is a "strong" advice to kernel to populate the page (and so avoid
page fault later) and it returns an error if kernel haven't succeed to
populate correctly.
This is made when application *want* the pages read. Not for performance
hint to boost next read.

We already have the `populate()` and `will_need()` advice/option.