refactor!: Remove `ErrorKind::WrongType`.
As we remove the error kind, we need to return an Option when creating
the builder property from the layout.
The simpler is to define a trait `FromLayoutProperty`.

As we do so, we also define `Property::as_builder` which avoid user
to avoid explicit type with `SpecificProperty::from_property(...)`.
They can use `property.as_builder()` and let's rust infer the type.