refactor!: VariantName and PropertyName don't have `ToString` as supertrait
`ToString` is mainly use to "display" a value.
(And clippy complain about that and propose to implement `Display` instead)

We don't want to display it but get a name (as a `[u8]`) to store in the
container.
Let's be clean on this and add `as_bytes` to `VariantName` and
`PropertyName` trait.