feature: Introduce `VariantIdBuilder` to build the correct variant kind.
`VariantIdProperty` allow to parse the variant id property in the entry
to get its raw identifier (an integer).
However, the order in which variants are declared may change between
creators. So we must not parse variant's properties based only on this
integer.

`VariantIdBuilder` is using the declared names of the variant in the
layout to properly map the variant id of the entry to the expected
variant in the code.